requirements.txt
pandas
numpy
scikit-learn
joblib
matplotlib
seaborn
fastapi
uvicorn
pydantic
python-dotenv
